# Conflicts:
#	src/main/java/com/example/lit/service/User/UserService.java
#	src/main/java/com/example/lit/service/User/UserServiceImple.java
#	src/main/resources/static/assets/js/admin/user-management.js

+++ src/main/java/com/example/lit/service/review/LitUpService.java
@@ -12,7 +12,7 @@ public interface LitUpService {
     // 좋아요 취소(작성자)
     public void removeLike(LikeVO likeVO);
     // 좋아요 갯수 카운트
-    public int getLikeTotal(Long reviewNumber);
+    public Long getLikeTotal(Long reviewNumber);
 
 
     //댓글 추가
